Refactor: Move auth callback to dynamic route
